- Split helper classes into separate sub modules - Add description to extended gcode settings - Fix bug with determining extended gcode mode when no entry exists for the current gcode - Remove numeric index support from Position and use named properties instead
